requirements.txt
# Qwen2.5 Inference Optimization & Deployment Course
# Complete Requirements File
#
# Install all dependencies:
#   pip install -r requirements.txt
#
# Note: PyTorch should be installed separately with CUDA support:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# Core LLM Inference
vllm>=0.6.0
transformers>=4.40.0
huggingface-hub>=0.20.0

# Performance Monitoring & Profiling
psutil>=5.9.0
gpustat>=1.1.0
torch-tb-profiler>=0.4.0
tensorboard>=2.13.0

# Data Analysis & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress & Utilities
tqdm>=4.65.0

# Load Testing (Week 7)
locust>=2.15.0
requests>=2.31.0

# Deployment & Cloud (Week 6)
# AWS CLI (install separately): pip install awscli
# Docker (system installation required)
# kubectl (system installation required)
# eksctl (system installation required)

# Development & Testing
pytest>=7.4.0  # Optional, for testing

# Jupyter Notebook (Optional, for interactive exploration)
jupyter>=1.0.0
ipykernel>=6.25.0

